nilm_predictor/requirements.txt
torch==2.0.1
psycopg2-binary==2.9.6
pandas==2.0.3
scikit-learn==1.0.2
numpy==1.24.4